Reorg the bundle
- make a `build` directory
- copy static files that need to go along with the build
- move webpack config out of `webpack` directory and into the root
