+++ package.json
@@ -4,23 +4,26 @@
   "version": "3.2.90",
   "description": "Truffle - Simple development framework for Ethereum",
   "dependencies": {
+    "mocha": "^3.4.2",
     "original-require": "^1.0.1",
     "solc": "^0.4.11"
   },
   "devDependencies": {
+    "clean-webpack-plugin": "^0.1.16",
+    "copy-webpack-plugin": "^4.0.1",
     "js-scrypt": "^0.2.0",
     "on-build-webpack": "^0.1.0",
     "prepend-file": "^1.3.1",
     "shebang-loader": "0.0.1",
     "stream-buffers": "^3.0.1",
-    "truffle-core": "^3.2.6",
+    "truffle-core": "^3.2.7",
     "webpack": "^2.5.1"
   },
   "bin": {
     "truffle": "./cli.bundled.js"
   },
   "scripts": {
-    "build-cli": "webpack --config ./webpack/cli.webpack.config.js"
+    "build-cli": "webpack --config ./cli.webpack.config.js"
   },
   "repository": {
     "type": "git",
